metadata changes

+++ WPS-Process-Development-COBWEB/src/pillar/automaticvalidation/LaplacePhotoBlurCheck.java
@@ -39,12 +39,15 @@ import eu.cobwebproject.qa.automaticvalidation.BlurCheckRunnable; 	// interface
 public class LaplacePhotoBlurCheck extends AbstractAlgorithm{
 	static Logger LOG = Logger.getLogger(LaplacePhotoBlurCheck.class);
 	
+		
+
+	
 	public static final long MAX_IMG_SIZE = 6000000;
 	
 	private ArrayList<String> errors;
 	
 	/**
-	 * @author Sam Meek (unotts) and Seb Clarke (Environment Systems)
+	 * @author Sam Meek (unotts) and Seb Clarke (Environment Systems) and Julian Rosser (unotts)
 	 * Process to check whether a photo is blurry by using the functionality from cobweb-qa library by Michael K
 	 * Output is the metadata field "Obs_Usability" which is 1 for pass criteria and 0 for not passing
 	 * result is observations with 1 or 0
@@ -121,7 +124,7 @@ public class LaplacePhotoBlurCheck extends AbstractAlgorithm{
 		}
 		
 		//add DQ_Field		
-		resultTypeBuilder.add("Obs_Usability", Integer.class);
+		resultTypeBuilder.add("Obs_Usability", Double.class);
 		
 		// Build the result feature type
 		SimpleFeatureType typeF = resultTypeBuilder.buildFeatureType();
@@ -160,6 +163,9 @@ public class LaplacePhotoBlurCheck extends AbstractAlgorithm{
 				// Do the check and set usability accordingly
 				BlurCheckRunnable blurChecker = new BlurCheckAwt(original, threshold, false);
 				blurChecker.run();	// This should probably run on a thread! 
+		
+				//Usability could be more than 0 or 1 with a ratio.
+				//I.e. If photoVariance < threshold then DQ = photoVariance / threshold
 				resultFeatureBuilder.set("Obs_Usability", blurChecker.pass?1:0);
 				
 				// Build result
@@ -228,4 +234,6 @@ public class LaplacePhotoBlurCheck extends AbstractAlgorithm{
 		LOG.warn("image file size: " + String.valueOf(imgSize));
 		return imgSize > maxImageSize || imgSize == -1;
 	}
+	
+
  }
